% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achor2time.R
\name{achor2time}
\alias{achor2time}
\title{Convert a proxy record to the time domain using anchor points}
\usage{
achor2time(anchor_points = NULL, data = NULL, genplot = TRUE)
}
\arguments{
\item{anchor_points}{Anchor points made using the \code{\link{astro_anchor}} function or a matrix in which the first column is depth
and the second column is time.}

\item{data}{Data set which needs to be converted from the depth to time domain using set anchor points.
The data set should consist of a matrix with 2 column the first column should be depth
and the second column should be a proxy value.}

\item{genplot}{If  \code{genplot=TRUE} then 3 plots stacked on top of each other will be plotted.
Plot 1: the original dataset
Plot 2: the depth time plot
Plot 3: the dataset in the time domain}
}
\value{
The output is a matrix with 2 columns.
The first column is time.
The second column sedimentation proxy value.

If \code{genplot=TRUE} then 3 plots stacked on top of each other will be plotted.
Plot 1: the original dataset.
Plot 2: the depth time plot.
Plot 3: the dataset in the time domain.
}
\description{
Convert a proxy record to the time domain using anchor points made using the \code{\link{astro_anchor}} function.
}
\examples{
# Use the \code{age_model_zeeden} example anchor points of Zeeden et al., (2013) \cr
#to anchor the \code{grey} data set of Zeeden et al., (2013) in the time domain.

achored2time <- achor2time(anchor_points=age_model_zeeden,
data=grey,
genplot=TRUE)

}
