% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_superlet.R
\name{analyze_superlet}
\alias{analyze_superlet}
\title{Conduct the superlet transform on a time series or signal}
\usage{
analyze_superlet(
  data,
  upperPeriod,
  lowerPeriod,
  Nf,
  c1,
  o = NULL,
  mult = TRUE,
  order_scaling = "log2",
  order_alpha = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{Input data as a matrix or data frame. The first column must
represent depth or time, and the second column the signal or proxy record.}

\item{upperPeriod}{Maximum period to be analysed. Controls the lowest
analysed frequency.}

\item{lowerPeriod}{Minimum period to be analysed. Controls the highest
analysed frequency.}

\item{Nf}{Number of frequencies used to construct the superlet spectrum.}

\item{c1}{Base number of cycles of the Morlet wavelet. Acts as the
fundamental wavelet width.}

\item{o}{numeric vector of length two defining the minimum and
maximum superlet order.}

\item{mult}{Logical. If TRUE, the number of cycles increases
multiplicatively with superlet order. If FALSE, cycles increase
additively.}

\item{verbose}{Logical. If TRUE, print progress and interpolation
information.}

\item{freq.scale}{Character string defining how frequencies are sampled.
One of \code{"log2"}, \code{"linear"}, or \code{"sqrt"}.
Default is \code{"log2"}, following the original superlet formulation.}

\item{scaling}{Character string defining how the number of wavelet
cycles varies with scale. One of
\code{"log2"}, \code{"linear"}, \code{"sqrt"}, \code{"quadratic"},
or \code{"power"}.}

\item{alpha}{Numeric. Exponent used when \code{scaling = "power"}.
Values greater than one emphasize high frequency sharpening, whereas
values smaller than one emphasize low frequency sharpening.}
}
\value{
A list with class \code{"analyze.superlet"} containing
Power: time frequency power spectrum
dt: sampling interval after interpolation
dj: number of frequencies
Power.avg: average spectral power
Period: physical periods corresponding to frequencies
nc: number of columns in the input data
nr: number of frequency levels
axis.1: x axis values (time or depth)
axis.2: y axis values (log2 scaled periods)
c1: base number of wavelet cycles
o: numeric vector of length two defining the minimum and
maximum superlet order. If NULL, all frequencies are analysed using
order one.
freq.scale: scaling of the frequencies
order.mode: order.mode ("none", "log2", "linear", "sqrt", "power")
x: interpolated x values
y: interpolated signal values
}
\description{
Compute the superlet transform using a complex Morlet wavelet following
the approach of Moca et al. (2021). The superlet transform increases
time frequency resolution by combining multiple wavelet orders at each
frequency. Both the frequency sampling and the scaling of superlet order
with frequency can be explicitly controlled.
}
\examples{
\donttest{
## Example 1. Using the Total Solar Irradiance data set of Steinhilber et al. (2012)
TSI_sl <-
  analyze_superlet(
    data = TSI,
    upperPeriod = 8192,
    lowerPeriod = 16,
    Nf = 100,
    c1 = 2,
    o = c(1, 5),
    mult = TRUE,
    order_scaling = "log2",
    order_alpha = 1,
    verbose = FALSE
  )


## Example 2. Using the magnetic susceptibility data set of Pas et al. (2018)
mag_sl <-
  analyze_superlet(
    data = mag,
    upperPeriod = 254,
    lowerPeriod = 0.1,
    Nf = 100,
    c1 = 2,
    o = c(1, 5),
    mult = TRUE,
    order_scaling = "log2",
    order_alpha = 1,
    verbose = FALSE
  )


## Example 3. Using the greyscale data set of Zeeden et al. (2013)
grey_sl <-
  analyze_superlet(
    data = grey,
    upperPeriod = 256,
    lowerPeriod = 0.02,
    Nf = 100,
    c1 = 2,
    o = c(2,5),
    mult = TRUE,
    order_scaling = "log2",
    order_alpha = 1,
    verbose = FALSE)
}

}
\references{
Moca, V. V., Bârzan, H., Nagy-Dăbâcan, A., and Mureșan, R. C. (2021).
Time frequency super resolution with superlets.
Nature Communications, 12, 337.
\url{https://doi.org/10.1038/s41467-020-20539-9}
}
\author{
Adapted from MATLAB implementations by V. V. Moca et al. (2021)
}
