% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_wavelet.R
\name{analyze_wavelet}
\alias{analyze_wavelet}
\title{Computes the wavelet power spectrum of a time series/signal}
\usage{
analyze_wavelet(
  data = NULL,
  dj = 1/20,
  lowerPeriod = 2,
  upperPeriod = 1024,
  verbose = TRUE,
  omega_nr = 6
)
}
\arguments{
\item{data}{Input data, should be a matrix or data frame in which
the first column is depth or time and the second column is proxy record.}

\item{dj}{Spacing between successive scales \code{Default=1/200}.}

\item{lowerPeriod}{Lowest period to be analyzed \code{Default=2},
scaling is done using power 2 so for the best plotting results select a value to the power or 2.}

\item{upperPeriod}{Upper period to be analyzed \code{Default=1024},
scaling is done using power 2 so for the best plotting results select a value to the power or 2.}

\item{verbose}{Print text \code{Default=TRUE}.}

\item{omega_nr}{Number of cycles contained within the Morlet wavelet}
}
\value{
The output is a list (wavelet object) with the results of the continuous wavelet transform (CWT).
}
\description{
Compute the continuous wavelet transform (CWT) using a Morlet wavelet
}
\examples{
\donttest{
#Example 1. Using the Total Solar Irradiance data set of Steinhilver et al., (2012)
TSI_wt <-
 analyze_wavelet(
   data = TSI,
   dj = 1/200,
   lowerPeriod = 16,
   upperPeriod = 8192,
   verbose = TRUE,
   omega_nr = 6
 )


#Example 2. Using the magnetic susceptibility data set of Pas et al., (2018)
mag_wt <-
analyze_wavelet(
data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = TRUE,
omega_nr = 10
)

#Example 3. Using the greyscale data set of Zeeden et al., (2013)
grey_wt <-
 analyze_wavelet(
   data = grey,
   dj = 1/200,
   lowerPeriod = 0.02,
   upperPeriod = 256,
   verbose = TRUE,
   omega_nr = 8
 )

}
}
\references{
Angi Roesch and Harald Schmidbauer (2018). WaveletComp: Computational
Wavelet Analysis. R package version 1.1.
\url{https://CRAN.R-project.org/package=WaveletComp}

Gouhier TC, Grinsted A, Simko V (2021). R package biwavelet: Conduct Univariate and Bivariate Wavelet Analyses. (Version 0.20.21),
\url{https://github.com/tgouhier/biwavelet}

Torrence, C., and G. P. Compo. 1998. A Practical Guide to Wavelet Analysis.
Bulletin of the American Meteorological Society 79:61-78.
\url{https://paos.colorado.edu/research/wavelets/bams_79_01_0061.pdf}

Morlet, Jean, Georges Arens, Eliane Fourgeau, and Dominique Glard.
"Wave propagation and sampling theory—Part I: Complex signal and scattering in multilayered media.
" Geophysics 47, no. 2 (1982): 203-221.
\doi{<doi:10.1190/1.1441328>}

J. Morlet, G. Arens, E. Fourgeau, D. Giard;
Wave propagation and sampling theory; Part II, Sampling theory and complex waves.
 Geophysics 1982 47 (2): 222–236. \doi{<doi:10.1190/1.1441329>}
}
\author{
Code based on  ased on the \link[WaveletComp]{analyze.wavelet} function of the WaveletComp R package
and \link[biwavelet]{wt} function of the biwavelet R package which are based on the
wavelet MATLAB code written by Christopher Torrence and Gibert P. Compo.
}
