% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{anchor_points_Bisciaro_al}
\alias{anchor_points_Bisciaro_al}
\title{XRF records of the Bisciaro Fm}
\description{
data set consist of the tie points between the Bisciaro_al record
of Arts (2014) and the la2011 solution of laskar et al., (20111)\cr
}
\details{
The data set is a matrix with the 4 columns.
The first column is the depth/time of the al proxy record tie-points.
The second column is the time value of the la2011 astronomical solution tie-points.
The third column is the Al value of the a; tie-point.
The fourth column is the eccentricity value of the la2011 astronomical solution tie-point.
}
\references{
M.C.M. Arts, 2014,
Magnetostratigrpahy and geochemical analysis of the early Miocene Bisciaro Formation
in the Contessa Valley (Northern Italy). Unpublished Bsc. thesis \cr

Laskar, J., M. Gastineau, J. B. Delisle, A. Farrés, and A. Fienga (2011b),
Strong chaos induced by close encounters with Ceres and Vesta, Astron. Astrophys.,
532, L4,<doi:10.1051/0004-6361/201117504> \cr
}
