% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro_bchron.R
\name{astro_bchron}
\alias{astro_bchron}
\title{Use astrochronological input in a Bchronology run}
\usage{
astro_bchron(
  age_constraints = NULL,
  n_simulations = NULL,
  track_period_incl_sd = NULL,
  tracked_cycle_period = NULL,
  tracked_cycle_period_unc = NULL,
  tracked_cycle_period_unc_dist = "u",
  output = 1,
  run_multicore = FALSE
)
}
\arguments{
\item{age_constraints}{A data frame object with age constraints
The data frame should consist of 6 columns
column 1: ID code of the age constraint
column 2: Age of the age constraint
column 3: Standard deviation in age (should be the SI number as the age)
column 4: Position of the age constraints in the depth domain
column 5: Thickness of the interval from which the the age constraints
originate from
column 6: calCurve values assign as "G" for a Gaussian distribution}

\item{n_simulations}{number of simulation runs}

\item{track_period_incl_sd}{A matrix of 3 columns in which the first column
is depth/height.The second column is the period of the tracked cycle.
The thirds column is uncertainty given as 1 standard deviation for the
period of the tracked cycle}

\item{tracked_cycle_period}{period in time of the tracked cycle}

\item{tracked_cycle_period_unc}{uncertainty in the period of the tracked cycle}

\item{tracked_cycle_period_unc_dist}{distribution of the uncertainty of the
tracked cycle value need to be either "u" for uniform distribution or
"n" for normal distribution  \code{Default="u"}}

\item{output}{If output = 1 a matrix with the predicted ages given the input
If output = 2 a matrix with 5 columns is generated, the first column is
depth/height, the second column are the quantile
(0.025,0.373,0.5,0.6827,0.975) age values of the runs}

\item{run_multicore}{Run function using multiple cores \code{Default="FALSE"}}
}
\value{
If output = 1 a matrix with the predicted ages given the input for each run
If output = 2 a matrix with 6 columns is generated, the first column is
depth/height, the other columns are the quantile
(0.025,0.373,0.5,0.6827,0.975) age values of the runs
}
\description{
Run a modified version of the Bchronology algorithm from
the R package "Bchron and modifiedBChron " (Haslett and Parnell, 2008) and (Trayler et al., 2020)
}
\examples{
\donttest{

# U/Pb ages of Kasbohm et al., (2021)
id <- c("CCT18_322", "CCT18_315", "CCT18_311")
ages <- c(20158, 20575, 20857)
ageSds <- c(28, 40, 34)
position <- c(10.2, 6.1, 3.2)
thickness <- c(0.2, 0.1, 0.1)
calCurves <- rep("G", length(ages))
ash_Bisc <- as.data.frame(cbind(id, ages, ageSds, position, thickness, calCurves))


Bisciaro_al <- Bisciaro_XRF[, c(1, 61)]
Bisciaro_al <-
 astrochron::sortNave(Bisciaro_al, verbose = FALSE, genplot = FALSE)
Bisciaro_al <-
 astrochron::linterp(Bisciaro_al,
                     dt = 0.01,
                     verbose = FALSE,
                     genplot = FALSE)
Bisciaro_al <- Bisciaro_al[Bisciaro_al[, 1] > 2.8,]

Bisciaro_al_wt <-
 analyze_wavelet(
   data = Bisciaro_al,
   dj = 1 / 200 ,
   lowerPeriod = 0.01,
   upperPeriod = 50,
   verbose = FALSE,
   omega_nr = 8
 )


Bisciaro_ca <- Bisciaro_XRF[, c(1, 55)]
Bisciaro_ca <-
 astrochron::sortNave(Bisciaro_ca, verbose = FALSE, genplot = FALSE)
Bisciaro_ca <-
 astrochron::linterp(Bisciaro_ca,
                     dt = 0.01,
                     verbose = FALSE,
                     genplot = FALSE)
Bisciaro_ca <- Bisciaro_ca[Bisciaro_ca[, 1] > 2.8,]

Bisciaro_ca_wt <-
 analyze_wavelet(
   data = Bisciaro_ca,
   dj = 1 / 200 ,
   lowerPeriod = 0.01,
   upperPeriod = 50,
   verbose = FALSE,
   omega_nr = 8
 )

Bisciaro_sial <- Bisciaro_XRF[, c(1, 64)]
Bisciaro_sial <-
 astrochron::sortNave(Bisciaro_sial, verbose = FALSE, genplot = FALSE)
Bisciaro_sial <-
 astrochron::linterp(Bisciaro_sial,
                     dt = 0.01,
                     verbose = FALSE,
                     genplot = FALSE)
Bisciaro_sial <- Bisciaro_sial[Bisciaro_sial[, 1] > 2.8,]

Bisciaro_sial_wt <-
 analyze_wavelet(
   data = Bisciaro_sial,
   dj = 1 / 200 ,
   lowerPeriod = 0.01,
   upperPeriod = 50,
   verbose = FALSE,
   omega_nr = 8
 )


Bisciaro_Mn <- Bisciaro_XRF[, c(1, 46)]
Bisciaro_Mn <-
 astrochron::sortNave(Bisciaro_Mn, verbose = FALSE, genplot = FALSE)
Bisciaro_Mn <-
 astrochron::linterp(Bisciaro_Mn,
                     dt = 0.01,
                     verbose = FALSE,
                     genplot = FALSE)
Bisciaro_Mn <- Bisciaro_Mn[Bisciaro_Mn[, 1] > 2.8,]

Bisciaro_Mn_wt <-
 analyze_wavelet(
   data = Bisciaro_Mn,
   dj = 1 / 200 ,
   lowerPeriod = 0.01,
   upperPeriod = 50,
   verbose = FALSE,
   omega_nr = 8
 )


Bisciaro_Mg <- Bisciaro_XRF[, c(1, 71)]
Bisciaro_Mg <-
 astrochron::sortNave(Bisciaro_Mg, verbose = FALSE, genplot = FALSE)
Bisciaro_Mg <-astrochron::linterp(Bisciaro_Mg,
                                 dt = 0.01,
                                 verbose = FALSE,
                                 genplot = FALSE)
Bisciaro_Mg <- Bisciaro_Mg[Bisciaro_Mg[, 1] > 2.8,]

Bisciaro_Mg_wt <-
 analyze_wavelet(
   data = Bisciaro_Mg,
   dj = 1 / 200 ,
   lowerPeriod = 0.01,
   upperPeriod = 50,
   verbose = FALSE,
   omega_nr = 8
 )

Bisciaro_ca_wt <-
 analyze_wavelet(
   data = Bisciaro_ca,
   dj = 1 / 200 ,
   lowerPeriod = 0.01,
   upperPeriod = 50,
   verbose = FALSE,
   omega_nr = 6
 )

wt_list_bisc <- list(Bisciaro_al_wt,
                    Bisciaro_ca_wt,
                    Bisciaro_sial_wt,
                    Bisciaro_Mn_wt,
                    Bisciaro_Mg_wt)

data_track_bisc <- cbind(
 Bisciaro_al_wt_track[, 2],
 Bisciaro_ca_wt_track[, 2],
 Bisciaro_sial_wt_track[, 2],
 Bisciaro_Mn_wt_track[, 2],
 Bisciaro_Mg_wt_track[, 2]
)

x_axis_bisc <- Bisciaro_al_wt_track[, 1]

retrack <- retrack_wt_MC(
 wt_list = wt_list_bisc,
 data_track = data_track_bisc,
 x_axis = x_axis_bisc,
 nr_simulations = 1000,
 seed_nr = 1337,
 verbose = TRUE,
 genplot = TRUE,
 keep_editable = FALSE,
 create_GIF = FALSE,
 plot_GIF = FALSE,
 width_plt =  600,
 height_plt = 450,
 period_up  =  1.5,
 period_down = 0.5,
 plot.COI = TRUE,
 n.levels = 100,
 palette_name = "rainbow",
 color_brewer = "grDevices",
 periodlab = "Period (metres)",
 x_lab = "depth (metres)",
 add_avg = FALSE,
 time_dir = TRUE,
 file_name = NULL,
 run_multicore = FALSE,
 output = 1,
 n_imgs = 50,
 plot_horizontal = TRUE,
 empty_folder = TRUE)

retrack <- retrack[[1]]


astro_bchron_res <- astro_bchron(age_constraints = ash_Bisc,
n_simulations = 10,
track_period_incl_sd = retrack,
tracked_cycle_period = 110,
tracked_cycle_period_unc = 21,
tracked_cycle_period_unc_dist = "u",
output=1,
run_multicore = FALSE)}


}
\references{
Jennifer Kasbohm, Blair Schoene, Alessandro Montanari,
Rodolfo Coccioni,High-precision U-Pb zircon geochronology of the Miocene
Bisciaro Formation, Contessa Section, Italy: A case study for requisite
radioisotopic calibration of bio- and magnetostratigraphy, Palaeogeography,
Palaeoclimatology, Palaeoecology, Volume 576, 2021, 110487, ISSN 0031-0182,
<\doi{doi:10.1016/j.palaeo.2021.110487}>
}
