% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{astrosignal_example}
\alias{astrosignal_example}
\title{An ETP astronomical solution}
\description{
The \code{\link{astrosignal_example}} is a pre-generated ETP \cr
(eccentricity-tilt-precession) (p-0.5t based on the la2004 solution) \cr
the \code{\link{astrosignal_example}} can be used to anchor the \code{\link{grey}}\cr
data set to an astronomical solution eg. \code{\link{astrosignal_example}} \cr
using the  \code{\link{astro_anchor}} function. the data set was generated using the \cr
\link[astrochron]{etp} function of the 'astrochron' R package.
 The pre-generated ETP spans 5000 to 6000kyr.

#'@details
Column 1: time (kyr)\cr
Column 2: ETP  \cr
}
\references{
Stephen R. Meyers,Cyclostratigraphy and the problem of astrochronologic testing,
Earth-Science Reviews,Volume 190,2019,Pages 190-223,ISSN 0012-8252
<\doi{doi:10.1016/j.earscirev.2018.11.015}>

J. Laskar, P. Robutel, F. Joutel, M. Gastineau, A.C.M. Correia, and B. Levrard, B., 2004,
A long term numerical solution for the insolation quantities of the Earth: Astron. Astrophys.,
Volume 428, 261-285. <\doi{doi:10.1051/0004-6361:20041335}>
}
\author{
Generated using the \link[astrochron]{etp}
function of the \link[astrochron]{astrochron-package}.
}
