% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{Bisciaro_sial_wt_track}
\alias{Bisciaro_sial_wt_track}
\title{Period of the short kyr ecc cycle in the si/Al record of the Bisciaro Fm}
\description{
Data points which give the period (in meters) of the short kyr eccentricity cycle tracked \cr
in the wavelet scalogram of the silicon/aluminium (XRF) record of the Bisciaro Formation \cr
The period was tracked using the \code{\link{track_period_wavelet}} function\cr
The tracking is based on a reinterpretation of Arts (2014)\cr

#'@details
Column 1: depth proxy record\cr
Column 2: period tracked in the wavelet scalogram of the silicon/aluminium (XRF) record
}
\references{
M.C.M. Arts, 2014,
Magnetostratigrpahy and geochemical analysis of the early Miocene Bisciaro Formation
in the Contessa Valley (Northern Italy). Unpublished Bsc. thesis \cr
}
