% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{Bisciaro_XRF}
\alias{Bisciaro_XRF}
\title{XRF records of the Bisciaro Fm}
\description{
XRF proxy records from the early Miocene Bisciaro Formation
in the Contessa Valley (Northern Italy)
}
\details{
Column 1: depth proxy record\cr
Column 2-71: XRF proxy records
}
\references{
M.C.M. Arts, 2014,
Magnetostratigrpahy and geochemical analysis of the early Miocene Bisciaro Formation
in the Contessa Valley (Northern Italy). Unpublished Bsc. thesis \cr
}
