% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{depth_rank_example}
\alias{depth_rank_example}
\title{An example depth rank series}
\description{
The \code{\link{depth_rank_example}} example data set is a depth rank series which \cr
can be used as input for the \code{\link{lithlog_disc}} function which creates a \cr
discritzed record which can be used as input in the \code{\link{analyze_wavelet}}\cr
function
}
\details{
Column 1: depth (meters)\cr
Column 2: depth rank \cr
}
