% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expSedRamp.R
\name{expSedRamp}
\alias{expSedRamp}
\title{exponentially ramped  sedimentation rate model to convert time to stratigraphy}
\usage{
expSedRamp(data, sr_start, sr_end)
}
\arguments{
\item{data}{Input data set  should consist of a matrix with 2 columns with the first being time
and the second being the proxy}

\item{sr_start}{Initial sedimentation rate (in cm/kyr).}

\item{sr_end}{Final sedimentation rate (in cm/kyr).}
}
\value{
Returns a list which contains 10 elements
element 1: time series in the depth (m) domain
element 2: sedimentation rate curve
}
\description{
Apply a exponentially increasing increasing sedimentation rate model
 to convert time to stratigraphy.
}
\examples{

data_1 <- etp(
tmin = 0,
tmax = 4000,
dt = 1,
eWt = 1.5,
oWt = 0.75,
pWt = 1,
esinw = T,
solution = NULL,
standardize = T,
genplot = T,
verbose = T
)
data_ramped_ls <- expSedRamp(data_1, sr_start = 1, sr_end = 5)
}
\references{
Routines for astrochronologic testing, astronomical time scale construction, and
time series analysis <doi:10.1016/j.earscirev.2018.11.015>
}
\author{
Based on the \link[astrochron]{sedRamp}
function of the 'astrochron' R package.
}
