% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_signal_standard_deviation.R
\name{extract_signal_standard_deviation}
\alias{extract_signal_standard_deviation}
\title{Extract a signal  using standard deviation}
\usage{
extract_signal_standard_deviation(
  wavelet = NULL,
  tracked_cycle_curve = NULL,
  multi = 1,
  extract_cycle = NULL,
  tracked_cycle_period = NULL,
  add_mean = TRUE,
  tune = FALSE,
  genplot_uncertainty_wt = TRUE,
  genplot_extracted = TRUE
)
}
\arguments{
\item{wavelet}{Wavelet object created using the \code{\link{analyze_wavelet}} function.}

\item{tracked_cycle_curve}{Curve of the cycle tracked using the
\code{\link{track_period_wavelet}} function. Any input (matrix or dataframe)
 in which the first column is depth or time and the second column is period should work.}

\item{multi}{multiple of the standard deviation to be used as boundaries for the cycle extraction
\code{Default=1}.}

\item{extract_cycle}{Period of the cycle to be extracted.}

\item{tracked_cycle_period}{Period of the tracked cycle.}

\item{add_mean}{Add mean to the extracted cycle \code{Default=TRUE}.}

\item{tune}{Tune data set using the \code{Default=tracked_cycle_curve} curve \code{Default=FALSE}.}

\item{genplot_uncertainty_wt}{Generate a wavelet spectra plot with the tracked curve and its
analytical uncertainty based the Gabor uncertainty principle applied
continuous wavelet transform using a Morlet wavelet on superimposed on top of it.
In the plot the red curve and blue curves are the upper and lower bounds
based on the \code{multi} parameter which x-times the standard deviation of uncertainty.
The black curve is the \code{Default=tracked_cycle_curve} curve.}

\item{genplot_extracted}{Generates a plot with the data set and
the extracted cycle on top \code{Default=TRUE} of it.}
}
\value{
Signal extracted from the wavelet spectra.
Output is a matrix with the first column being depth/time
and the second column is the astronomical cycle extracted from the proxy record

If \code{genplot_uncertainty_wt=TRUE} then a wavelet spectra will be plotted
with the uncertainty superimposed on top of it. In the plot the red curve and
blue curves are the upper and lower bounds
based on the \code{multi} parameter.The black curve is the \code{Default=tracked_cycle_curve} curve.
If \code{genplot_extracted=TRUE} plot with the data set and
the extracted cycle on top of it will be plotted.
}
\description{
Extract signal from a
wavelet spectra in the depth domain using a the standard deviation of the omega (number of cycles)
as boundaries. The uncertainty is based on the Gabor uncertainty principle applied to the
continuous wavelet transform using a Morlet wavelet. The calculated uncertainty is the underlying
analytical uncertainty which is the result of applying the Gabor uncertainty principle to the
continuous wavelet transform using a Morlet wavelet.
}
\examples{
\donttest{
#Extract the 405 kyr eccentricity cycle from the the magnetic susceptibility \cr
#record of the Sullivan core and use the Gabor uncertainty principle to define \cr
# the mathematical uncertainty of the analysis and use a factor of that standard \cr
#  deviation to define boundaries

# perform the CWT
mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = TRUE,
omega_nr = 10)

#Track the 405 kyr eccentricity cycle in a wavelet spectra

#mag_track <- track_period_wavelet(astro_cycle = 405,
#                                   wavelet=mag_wt,
#                                   n.levels = 100,
#                                   periodlab = "Period (metres)",
#                                   x_lab = "depth (metres)")

#Instead of tracking, the tracked solution data set \code{\link{mag_track_solution}} is used \cr
mag_track <- mag_track_solution

mag_track_complete <- completed_series(
  wavelet = mag_wt,
  tracked_curve = mag_track,
  period_up = 1.2,
  period_down = 0.8,
  extrapolate = TRUE,
  genplot = TRUE
)

# smooth the tracking of the 405 kyr eccentricity cycle
mag_track_complete <- loess_auto(time_series = mag_track_complete,
genplot = TRUE, print_span = TRUE)

# extract the 405 kyr eccentricty cycle from the wavelet spectrum and use \cr
# the Gabor uncertainty principle to define the mathematical uncertainty of \cr
# the analysis and use a multiple of the derived standard deviation to define boundaries

mag_405_ecc <- extract_signal_standard_deviation(
wavelet = mag_wt,
tracked_cycle_curve = mag_track_complete,
multi = 1,
extract_cycle = 405,
tracked_cycle_period = 405,
add_mean = TRUE,
tune = FALSE,
genplot_uncertainty_wt = TRUE,
genplot_extracted = TRUE
)
}
}
\references{
Gabor, Dennis. "Theory of communication. Part 1: The analysis of information."
Journal of the Institution of Electrical Engineers-part III: radio and
communication engineering 93, no. 26 (1946): 429-441.

Russell, Brian, and Jiajun Han. "Jean Morlet and the continuous wavelet transform.
" CREWES Res. Rep 28 (2016): 115.
}
