% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{grey_track}
\alias{grey_track}
\title{Tracking points of the precession (22 kyr cycle) IODP 926 grey scale (154-174m) record of Zeeden et al., (2013)}
\description{
Example data which consists of tracking points of the precession (22 kyr cycle) in the wavelet
spectra of the IODP 926 grey scale (154-174m) record of Zeeden et al., (2013)
}
\details{
Column 1: Depth (meters)\cr
Column 2: period (meters)\cr
}
\references{
Christian Zeeden, Frederik Hilgen, Thomas Westerhold, Lucas Lourens, Ursula Röhl, Torsten Bickert,
Revised Miocene splice, astronomical tuning and calcareous plankton biochronology of ODP Site 926 between 5 and 14.4Ma,
Palaeogeography, Palaeoclimatology, Palaeoecology,Volume 369,2013,Pages 430-451,ISSN 0031-0182,
<doi:10.1016/j.palaeo.2012.11.009>
}
