% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{GTS_info}
\alias{GTS_info}
\title{Information of the Geological timescale 2020}
\description{
GTS_info data set consists the information of the Geological
timescale 2020 including the color data of Ogg et al., (2021)
The ages, durations, uncertainties and colors of the Geological
timescale 2020 are included in the data set
}
\details{
Column 1: name	\cr
Column 2: type	\cr
Column 1: top age\cr
Column 1: top error\cr
Column 1: bottom age\cr
Column 1: bottom error\cr
Column 1: Cyan value\cr
Column 1: Magenta value\cr
Column 1: Yellow value\cr
Column 1: Key  value\cr
Column 1: Red Value\cr
Column 1: Green value\cr
Column 1: Blue value	\cr
Column 1: font style	\cr
Column 1: font color\cr
}
\references{
Ogg, Gabi & Ogg, James & Gradstein, Felix. (2021).
Recommended color coding of stages - Appendix 1
from Geologic Time Scale 2020.
}
