% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{mag}
\alias{mag}
\title{Magnetic susceptibility data of the Sullivan core of Pas et al., (2018)}
\description{
The magnetic susceptibility data set consists
of the magnetic susceptibility measurements of Pas et al., (2018), which measured the magnetic
 susceptibility on the Sullivan core which is of Famennian age.
}
\details{
Column 1: depth value (meters depoth)\cr
Column 2: magnetic susceptibility  value\cr
}
\references{
Damien Pas, Linda Hinnov, James E. (Jed) Day, Kenneth Kodama, Matthias Sinnesael, Wei Liu,
Cyclostratigraphic calibration of the Famennian stage (Late Devonian, Illinois Basin, USA),
Earth and Planetary Science Letters,
Volume 488,
2018,
Pages 102-114,
ISSN 0012-821X,
\doi{<doi:1016/j.epsl.2018.02.010>}
}
