% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{mag_track_solution}
\alias{mag_track_solution}
\title{Period of the 405 kyr ecc cycle in the magnetic susceptibility record of the Sullivan core}
\description{
Data points which give the period (in meters) of the 405 kyr eccentricity cycle tracked
in the wavelet spectra of the magnetic susceptibility record of the Sullivan core \cr
The period was tracked using the \code{\link{track_period_wavelet}} function\cr
The tracking is based on the original age model of Pas et al., (2018)\cr
}
\details{
Column 1: Depth (meters)\cr
Column 2: tracked period of 405 kyr eccentricity cycle (meters)\cr
}
\references{
Damien Pas, Linda Hinnov, James E. (Jed) Day, Kenneth Kodama, Matthias Sinnesael, Wei Liu,
Cyclostratigraphic calibration of the Famennian stage (Late Devonian, Illinois Basin, USA),
Earth and Planetary Science Letters,
Volume 488,
2018,
Pages 102-114,
ISSN 0012-821X,
\doi{<doi:10.1016/j.epsl.2018.02.010>}
}
