% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_red_noise_wt.R
\name{model_red_noise_wt}
\alias{model_red_noise_wt}
\title{Models average spectral power based curves based on a red-noise signal
generated using the characteristics of an input signal.}
\usage{
model_red_noise_wt(wavelet = NULL, n_simulations = NULL, verbose = TRUE)
}
\arguments{
\item{wavelet}{Wavelet object created using the \code{\link{analyze_wavelet}} function.}

\item{n_simulations}{Number of red noise simulations.}

\item{verbose}{Print text \code{Default=TRUE}.}
}
\value{
Returns a matrix in which each column represents the average spectral
power resulting from a red-noise run.
}
\description{
The \code{\link{model_red_noise_wt}} function is used to generate
average spectral power curves based on and input signal and set wavelet settings.
}
\examples{
\dontrun{
#'#generate average spectral power curves based on red noise curves which are\cr
# based on the magnetic susceptibility record of the Sullivan core of De pas et al., (2018)

mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = FALSE,
omega_nr = 10)

#increase n_simulations to better define the red noise spectral power curve
mag_wt_red_noise <- model_red_noise_wt(wavelet=mag_wt,
n_simulations=100,
verbose=FALSE)
}


}
