% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_astro_anchor.R
\name{plot_astro_anchor}
\alias{plot_astro_anchor}
\title{Plot proxy record anchored to an astronomical solution}
\usage{
plot_astro_anchor(
  astro_solution = NULL,
  proxy_signal = NULL,
  anchor_points = NULL,
  time_dir = TRUE,
  keep_editable = FALSE
)
}
\arguments{
\item{astro_solution}{Input is an astronomical solution with with the the proxy record was be anchored to,
the input should be a matrix or data frame with the first column being
age and the second column should be a insolation/angle/value}

\item{proxy_signal}{Input is the proxy data set which will which was
anchored to an astronomical solution, the input should be a matrix or
data frame with the first column being  depth/time and the second column should be a proxy value.}

\item{anchor_points}{anchor points generated using the \code{\link{astro_anchor}} function}

\item{time_dir}{The direction of the proxy record which was assumed during achoring if time increases with increasing depth/time values
(e.g. bore hole data which gets older with increasing depth ) then time_dir should be set to TRUE
if time decreases with depth/time values (eg stratospheric logs where 0m is the bottom of the section)
then time_dir should be set to FALSE \code{time_dir=TRUE}}

\item{keep_editable}{Keep option to add extra features after plotting  \code{Default=FALSE}}
}
\value{
The output is a set of 2 plots connected by lines
The top plot is the proxy record with anchor points on top of it
The bottom plot is the astronomical solution
The lines connect the anchro points
}
\description{
Plot the results of the anchoring  the extracted signal to an astronomical solution using
The \code{\link{astro_anchor}} f
}
\examples{
\donttest{
# Use the grey_track example tracking points to anchor the grey scale data set
# of Zeeden et al., (2013) to the p-0.5t la2004 solution

grey_wt <-
 analyze_wavelet(
   data = grey,
   dj = 1/200,
   lowerPeriod = 0.02,
   upperPeriod = 256,
   verbose = FALSE,
   omega_nr = 8
 )

#Use the pretracked grey_track curve which traced the precession cycle
grey_track <- completed_series(
 wavelet = grey_wt,
 tracked_curve  = grey_track,
 period_up  = 1.25,
 period_down  = 0.75,
 extrapolate = TRUE,
genplot = FALSE
)
# Extract precession, obliquity and eccentricity to create a synthetic insolation curve

grey_prec <- extract_signal(
tracked_cycle_curve = grey_track[,c(1,2)],
wavelet = grey_wt,
period_up = 1.2,
period_down = 0.8,
add_mean = FALSE,
tracked_cycle_period = 22,
extract_cycle = 22,
tune = FALSE,
plot_residual = FALSE
)

grey_obl <- extract_signal(
 tracked_cycle_curve = grey_track[,c(1,2)],
 wavelet = grey_wt,
 period_up = 1.2,
 period_down = 0.8,
 add_mean = FALSE,
 tracked_cycle_period = 22,
 extract_cycle = 110,
 tune = FALSE,
 plot_residual = FALSE
)

grey_ecc <- extract_signal(
 tracked_cycle_curve = grey_track[,c(1,2)],
 wavelet = grey_wt,
 period_up = 1.25,
 period_down = 0.75,
 add_mean = FALSE,
 tracked_cycle_period = 22,
 extract_cycle = 40.8,
 tune = FALSE,
 plot_residual = FALSE
)

insolation_extract <- cbind(grey_ecc[,1],grey_prec[,2]+grey_obl[,2]+grey_ecc[,2]+mean(grey[,2]))
insolation_extract <- as.data.frame(insolation_extract)
insolation_extract_mins <- min_detect(insolation_extract,pts=3)

#use the astrosignal_example to tune to which is an \cr
# ETP solution (p-0.5t la2004 solution)
astrosignal_example <- na.omit(astrosignal_example)
astrosignal_example[,2] <- -1*astrosignal_example[,2]
astrosignal <- as.data.frame(astrosignal_example)

#'#anchor the synthetic insolation curve extracted from the greyscale record to the insolation curve.
#'#use the anchor_points_grey dataset to plot a result of using the anchor_points function

#'anchor_points <- astro_anchor(
#'astro_solution = astrosignal,
#'proxy_signal = insolation_extract,
#'proxy_min_or_max = "min",
#'clip_astrosolution = FALSE,
#'astrosolution_min_or_max = "min",
#'clip_high = NULL,
#'clip_low = NULL,
#'extract_astrosolution  = FALSE,
#'astro_period_up  = NULL,
#'astro_period_down  = NULL,
#'astro_period_cycle  = NULL,
#'extract_proxy_signal  = FALSE,
#'proxy_period_up  = NULL,
#'proxy_period_down  = NULL,
#'proxy_period_cycle  = NULL,
#'pts=3,
#'verbose=FALSE,
#'genplot=FALSE #  set verbose to TRUE to allow for anchoring using text feedback commands
)


plot_astro_anchor(astro_solution = astrosignal,
proxy_signal = insolation_extract,
anchor_points = anchor_points_grey,
time_dir =TRUE,
keep_editable = FALSE)


)}

}
