% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eha_log2.R
\name{plot_eha_log2}
\alias{plot_eha_log2}
\title{Plots an log2 spaced EHA spectra}
\usage{
plot_eha_log2(
  eha_log2 = NULL,
  plot_opt = "Amplitude",
  lowerPeriod = NULL,
  upperPeriod = NULL,
  n.levels = 100,
  palette_name = "rainbow",
  color_brewer = "grDevices",
  useRaster = TRUE,
  periodlab = "Period (metres)",
  x_lab = "depth (metres)",
  keep_editable = FALSE,
  dev_new = TRUE,
  plot_dir = TRUE,
  add_lines = NULL,
  add_points = NULL,
  add_abline_h = NULL,
  add_abline_v = NULL,
  add_MTM_peaks = FALSE,
  add_data = TRUE,
  add_avg = FALSE,
  pval_abline = c(0.1, 0.05),
  pval_cutoff = c(0.1),
  add_MTM = FALSE,
  mtm_siglvl = 0.95,
  demean_mtm = TRUE,
  detrend_mtm = TRUE,
  padfac_mtm = 5,
  tbw_mtm = 3,
  plot_horizontal = TRUE
)
}
\arguments{
\item{eha_log2}{eha_log2 object created using the \code{\link{eha_log2}} function.}

\item{plot_opt}{plot options are  "Power", "Amplitude", "Probability" or "F_test"}

\item{lowerPeriod}{Lowest period value which will be plotted}

\item{upperPeriod}{Highest period value which will be plotted}

\item{n.levels}{Number of color levels \code{Default=100}.}

\item{palette_name}{Name of the color palette which is used for plotting.
The color palettes than can be chosen depends on which the R package is specified in
the color_brewer parameter. The included R packages from which palettes can be chosen
from are; the 'RColorBrewer', 'grDevices', 'ColorRamps' and 'Viridis' R packages.
There are many options to choose from so please
read the documentation of these packages \code{Default=rainbow}.
The R package 'viridis' has the color palette options: “magma”, “plasma”,
“inferno”, “viridis”, “mako”, and “rocket”  and “turbo”
To see the color palette options of the The R pacakge 'RColorBrewer' run
the RColorBrewer::brewer.pal.info() function
The R package 'colorRamps' has the color palette options:"blue2green",
"blue2green2red", "blue2red",    "blue2yellow", "colorRamps",    "cyan2yellow",
"green2red", "magenta2green", "matlab.like", "matlab.like2" and    "ygobb"
The R package 'grDevices' has the built in  palette options:"rainbow",
"heat.colors", "terrain.colors","topo.colors" and "cm.colors"
To see even more color palette options of the The R pacakge 'grDevices' run
the grDevices::hcl.pals() function}

\item{color_brewer}{Name of the R package from which the color palette is chosen from.
The included R packages from which palettes can be chosen
are; the RColorBrewer, grDevices, ColorRamps and Viridis R packages.
There are many options to choose from so please
read the documentation of these packages. "\code{Default=grDevices}}

\item{useRaster}{Plot as a raster or vector image \code{Default=TRUE}.
WARNING plotting as a vector image is computationally intensive.}

\item{periodlab}{Label for the y-axis \code{Default="Period (metres)"}.}

\item{x_lab}{Label for the x-axis \code{Default="depth (metres)"}.}

\item{keep_editable}{Keep option to add extra features after plotting  \code{Default=FALSE}}

\item{dev_new}{Opens a new plotting window to plot the plot, this guarantees a "nice" looking plot however when plotting in an R markdown
document the plot might not plot  \code{Default=TRUE}}

\item{plot_dir}{The direction of the proxy record which is assumed for tuning if time increases with increasing depth/time values
(e.g. bore hole data which gets older with increasing depth ) then plot_dir should be set to TRUE
if time decreases with depth/time values (eg stratospheric logs where 0m is the bottom of the section)
then plot_dir should be set to FALSE \code{plot_dir=TRUE}}

\item{add_lines}{Add  lines to the wavelet plot input should be matrix with first axis being depth/time the columns after that
should be period values  \code{Default=NULL}}

\item{add_points}{Add points to the wavelet plot input should be matrix with first axis being depth/time and columns after that
should be period values \code{Default=NULL}}

\item{add_abline_h}{Add horizontal lines to the plot. Specify the lines as a vector e.g. c(2,3,5,6)  \code{Default=NULL}}

\item{add_abline_v}{Add vertical lines to the plot. Specify the lines as a vector e.g. c(2,3,5,6)  \code{Default=NULL}}

\item{add_MTM_peaks}{Add the MTM peak periods as horizontal lines \code{Default=FALSE}}

\item{add_data}{Plot the data on top of the wavelet \code{Default=TRUE}}

\item{add_avg}{Plot the average wavelet spectral power to the side of the wavelet \code{Default=FALSE}}

\item{add_MTM}{Add the MTM  plot next to the wavelet plot \code{Default=FALSE}}

\item{mtm_siglvl}{select the significance level  (0-1) for the MTM spectrum \code{Default=0.95}}

\item{demean_mtm}{Remove mean from data before conducting the MTM analysis \code{Default=TRUE}}

\item{detrend_mtm}{Remove mean from data before conducting the MTM analysis \code{Default=TRUE}}

\item{padfac_mtm}{Pad factor for the MTM analysis \code{Default=5}}

\item{tbw_mtm}{time bandwidth product of the MTM analysis  \code{Default=3}}

\item{plot_horizontal}{plot the wavelet horizontal or vertical eg y axis is depth or y axis power \code{Default=TRUE}}
}
\value{
The output is a plot of a EHA spectra.
if add_MTM_peaks = TRUE then the output of the MTM analysis will given as matrix
}
\description{
Plot wavelet scalogram using the outcome of the \code{\link{eha_log2}} function.
}
\examples{
\donttest{
#Example 1. A plot of a wavelet spectra using the Total Solar Irradiance
# data set of Steinhilber et al., (2012)

TSI_eha_log2 <- eha_log2(data = TSI,
win =  8192,
tbw = 4,
demean = TRUE,
detrend = TRUE,
upperPeriod = 8192,
lowerPeriod = c,
pad = NULL,
padding = "noise")

plot_eha_log2(
eha_log2 = TSI_eha_log2,
plot_opt = "Amplitude",
lowerPeriod = 16,
upperPeriod = 8192,
n.levels = 100,
palette_name = "rainbow",
color_brewer = "grDevices",
useRaster = TRUE,
periodlab = "Period (metres)",
x_lab = "depth (metres)",
keep_editable = FALSE,
dev_new = TRUE,
plot_dir = TRUE,
add_lines = NULL,
add_points = NULL,
add_abline_h = NULL,
add_abline_v = NULL,
add_MTM_peaks = FALSE,
add_data = TRUE,
add_avg = FALSE,
add_MTM = FALSE,
mtm_siglvl = 0.95,
demean_mtm = TRUE,
detrend_mtm = TRUE,
padfac_mtm = 5,
tbw_mtm = 3,
plot_horizontal = TRUE)

#Example 2. A plot of a wavelet spectra using the magnetic susceptibility
#data set of Pas et al., (2018)
mag_eha_log2 <- eha_log2(data = mag,
win =  254,
tbw = 4,
demean = TRUE,
detrend = TRUE,
upperPeriod = 254,
lowerPeriod = 0.1,
pad = NULL,
padding = "noise")

plot_eha_log2(
eha_log2 = mag_eha_log2,
plot_opt = "Amplitude",
lowerPeriod = 0.1,
upperPeriod = 254,
n.levels = 100,
palette_name = "rainbow",
color_brewer = "grDevices",
useRaster = TRUE,
periodlab = "Period (metres)",
x_lab = "depth (metres)",
keep_editable = FALSE,
dev_new = TRUE,
plot_dir = TRUE,
add_lines = NULL,
add_points = NULL,
add_abline_h = NULL,
add_abline_v = NULL,
add_MTM_peaks = FALSE,
add_data = TRUE,
add_avg = FALSE,
add_MTM = FALSE,
mtm_siglvl = 0.95,
demean_mtm = TRUE,
detrend_mtm = TRUE,
padfac_mtm = 5,
tbw_mtm = 3,
plot_horizontal = TRUE)


#Example 3. A plot of a wavelet spectra using the greyscale
# data set of Zeeden et al., (2013)
grey_eha_log2 <- eha_log2(data = grey,
win =  256,
tbw = 4,
demean = TRUE,
detrend = TRUE,
upperPeriod = 256,
lowerPeriod = 0.02,
pad = NULL,
padding = "noise")


plot_eha_log2(
eha_log2 = grey_eha_log2,
plot_opt = "Amplitude",
lowerPeriod = 0.02,
upperPeriod = 256,
n.levels = 100,
palette_name = "rainbow",
color_brewer = "grDevices",
useRaster = TRUE,
periodlab = "Period (metres)",
x_lab = "depth (metres)",
keep_editable = FALSE,
dev_new = TRUE,
plot_dir = TRUE,
add_lines = NULL,
add_points = NULL,
add_abline_h = NULL,
add_abline_v = NULL,
add_MTM_peaks = FALSE,
add_data = TRUE,
add_avg = FALSE,
add_MTM = FALSE,
mtm_siglvl = 0.95,
demean_mtm = TRUE,
detrend_mtm = TRUE,
padfac_mtm = 5,
tbw_mtm = 3,
plot_horizontal = TRUE)
}

}
\references{
Angi Roesch and Harald Schmidbauer (2018). WaveletComp: Computational
Wavelet Analysis. R package version 1.1.
\url{https://CRAN.R-project.org/package=WaveletComp}

S.R. Meyers, 2012, Seeing Red in Cyclic Stratigraphy: Spectral Noise Estimation for
Astrochronology: Paleoceanography, 27, PA3228, <doi:10.1029/2012PA002307>

S.R. Meyers, 2019 Cyclostratigraphy and the problem of astrochronologic testing, Earth-Science Reviews,
Volume 190, <doi.org/10.1016/j.earscirev.2018.11.015.>
}
\author{
Code based on the wt.image" functions of the 'WaveletComp' R package
The EHA and MTM analysis parts are from the astrochron R package of Meyers et al., (2012)
}
