% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_win_fft.R
\name{plot_win_fft}
\alias{plot_win_fft}
\title{plot windowed fft based spectral analysis results}
\usage{
plot_win_fft(
  win_fft = NULL,
  x_lab = c("depth (m)"),
  y_lab = c("frequency cycle/metre"),
  plot_res = 1,
  perc_vis = 0,
  freq_max = NULL,
  freq_min = NULL,
  keep_editable = FALSE
)
}
\arguments{
\item{win_fft}{list which is the results of the \code{\link{win_fft}}}

\item{x_lab}{label for the y-axis \code{Default="depth"}}

\item{y_lab}{label for the y-axis \code{Default="sedrate"}}

\item{plot_res}{plot 1 of 8 options option 1: Amplitude matrix,
option 2: Power matrix,
option 3: Phase matrix,
option 4: AR1_CL matrix,
option 5: AR1_Fit matrix ,
option 6: AR1_90_power matrix,
option 7: AR1_95_power matrix,
option 8: AR1_99_power matrix, \code{Default=1}}

\item{perc_vis}{Cutoff percentile when plotting \code{Default=0}}

\item{freq_max}{Maximum frequency to plot}

\item{freq_min}{Minimum frequency to plot}

\item{keep_editable}{Keep option to add extra features after plotting  \code{Default=FALSE}}
}
\value{
Returns a plot of, which plot 1 of 8 options,
option 1: Amplitude matrix
option 2: Power matrix
option 3: Phase matrix
option 4: AR1_CL matrix
option 5: AR1_Fit matrix
option 6: AR1_90_power matrix
option 7: AR1_95_power matrix
option 8: AR1_99_power matrix
}
\description{
The \code{\link{plot_win_fft}} function allows for the (re)plotting of the results of the \code{\link{win_fft}}
}
\examples{
\donttest{
#Conduct a windowed ftt on the magnetic susceptibility record \cr
# of the Sullivan core of Pas et al., (2018).

mag_win_fft <- win_fft(data= mag,
                   padfac = 5,
                   window_size = 12.5,
                   run_multicore = FALSE,
                   genplot = FALSE,
                   x_lab = c("depth (m)"),
                   y_lab = c("frequency cycle/metre"),
                   plot_res = 1,
                   perc_vis = 0.5,
                   freq_max = 5,
                   freq_min = 0.001,
                   keep_editable=FALSE)

# Plot the mplitude spectra
plot_win_fft(win_fft= mag_win_fft,
x_lab = c("depth (m)"),
y_lab = c("frequency cycle/metre"),
plot_res = 1,
perc_vis = 0.5,
freq_max = 5,
freq_min = 0.001,
keep_editable=FALSE)

}

}
