% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_power_sedrate.R
\name{sum_power_sedrate}
\alias{sum_power_sedrate}
\title{Calculate sum of maximum spectral power for sedimentation rates based
for a wavelet spectra}
\usage{
sum_power_sedrate(
  red_noise = NULL,
  wavelet = NULL,
  percentile = NULL,
  sedrate_low = NULL,
  sedrate_high = NULL,
  spacing = NULL,
  cycles = c(NULL),
  x_lab = "depth",
  y_lab = "sedrate",
  genplot = TRUE
)
}
\arguments{
\item{red_noise}{Red noise curves generated using the \code{\link{model_red_noise_wt}} function}

\item{wavelet}{Wavelet object created using the \code{\link{analyze_wavelet}} function}

\item{percentile}{Percentile value (0-1) of the rednoise runs which is used to normalize the data for.
To account for the distribution/distortion of the spectral power distribution based on the analytical technique and
random red-noise the data is normalized against a percentile based red-noise curve which is the results of the
'\code{\link{model_red_noise_wt}} modelling runs.}

\item{sedrate_low}{Minimum sedimentation rate (cm/kyr)for which the sum of maximum spectral power is calculated for.}

\item{sedrate_high}{Maximum sedimentation rate (cm/kyr) for which the sum of maximum spectral power is calculated  for.}

\item{spacing}{Spacing (cm/kyr) between sedimentation rates}

\item{cycles}{Astronomical cycles (in kyr) for which the combined sum of maximum spectral power is calculated for}

\item{x_lab}{label for the y-axis \code{Default="depth"}}

\item{y_lab}{label for the y-axis \code{Default="sedrate"}}

\item{genplot}{Generate plot \code{Default="TRUE"}}
}
\value{
Returns a matrix with sum of maximum spectral power for a given sedimentation rates and a given depths.

If \code{Default="TRUE"} a plot is created with 3 subplots.
Subplot 1 is plot in which the the sum of maximum spectral power for a
given sedimentation rate is plotted for each depth given depth.
Subplot 2 is a plot in which the average sum of maximum spectral power is plotted fro each sedimentation
Subplot 3 is a color scale for subplot 1.
}
\description{
The \code{\link{sum_power_sedrate}} function is used calculate the sum of
maximum spectral power for a list of astronomical cycles from a wavelet spectra.
The data is first normalized using the average spectral power curves
 for a given percentile based on results of the \code{\link{model_red_noise_wt}} function
}
\examples{
\dontrun{
#estimate sedimentation rate for the the magnetic susceptibility record \cr
# of the Sullivan core of De pas et al., (2018).

mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = FALSE,
omega_nr = 10)

#increase n_simulations to better define the red noise spectral power curve
mag_wt_red_noise <- model_red_noise_wt(wavelet=mag_wt,
n_simulations=100,
verbose=FALSE)

sedrates <- sum_power_sedrate(red_noise=mag_wt_red_noise,
wavelet=mag_wt,
percentile=0.75,
sedrate_low = 0.5,
sedrate_high = 4,
spacing = 0.05,
cycles = c(2376,1600,1180,696,406,110),
x_lab="depth",
y_lab="sedrate",
genplot = TRUE)
}


}
