% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_period_wavelet.R
\name{track_period_wavelet}
\alias{track_period_wavelet}
\title{Track the period of a cycle in a wavelet spectra}
\usage{
track_period_wavelet(
  astro_cycle = 405,
  wavelet = NULL,
  n.levels = 100,
  periodlab = "Period (metres)",
  x_lab = "depth (metres)"
)
}
\arguments{
\item{astro_cycle}{Duration (in kyr) of the cycle which traced.}

\item{wavelet}{Wavelet object created using the \code{\link{analyze_wavelet}} function.}

\item{n.levels}{Number of color levels \code{Default=100}.}

\item{periodlab}{label for the y-axis \code{Default="Period (metres)"}.}

\item{x_lab}{label for the x-axis \code{Default="depth (metres)"}.}
}
\value{
Results of the tracking of a cycle in the wavelet spectra is a matrix with 3 columns.
The first column is depth/time
The second column is the period of the tracked cycle
The third column is the sedimentation rate based on the duration (in time) of the tracked cycle
}
\description{
Interactively select points in a wavelet spectra to trace a period in a wavelet spectra.
The \code{\link{track_period_wavelet}} function plots a wavelet spectra in which spectral peaks can selected
allowing one to track a ridge hence one can track the a cycle with a changing period.
}
\examples{
\dontrun{
#Track the 405kyr eccentricity cycle in the magnetic susceptibility record \cr
# of the Sullivan core of De pas et al., (2018)

mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = TRUE,
omega_nr = 10)

mag_track <- track_period_wavelet(astro_cycle = 405,
                                  wavelet=mag_wt,
                                  n.levels = 100,
                                  periodlab = "Period (metres)",
                                  x_lab = "depth (metres)")
}

}
