% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{TSI}
\alias{TSI}
\title{Total solar irradiation data (0-9400ka) of steinhilber et al., (2012)}
\description{
The Total solar irradiation data set consists of the TSI values of steinhilber et al., (2012)
}
\details{
Column 1: Age (kyr)\cr
Column 2: Total solar Irradiation (TSI)\cr
}
\references{
Steinhilber, Friedhelm & Abreu, Jacksiel & Beer, Juerg & Brunner,
Irene & Christl, Marcus & Fischer, Hubertus & Heikkilä, U. & Kubik,
Peter & Mann, Mathias & Mccracken, K. & Miller, Heinrich & Miyahara,
Hiroko & Oerter, Hans & Wilhelms, Frank. (2012).
9,400 Years of cosmic radiation and solar activity from ice cores and tree rings.
Proceedings of the National Academy of Sciences of the United States of America.
109. 5967-71. 10.1073/pnas.1118965109.
\doi{<doi:10.1073/pnas.1118965109>}
}
