% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR.R
\name{WaverideR}
\alias{WaverideR}
\title{Extracting Signals from Wavelet Spectra}
\description{
The functions of the WaverideR package aid in the manipulation and extraction of
 cyclic signals using wavelets
}
\details{
Package: WaverideR

Type: R package

Version: 0.2.0 (begin of 2023)

License: GPL (>= 2)
}
\note{
If you want to use this package for publication or research
purposes, please cite: to be submitted
}
\author{
Michiel Arts

Maintainer: Michiel Arts \email{michiel.arts@stratigraphy.eu}
}
