% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaverideR_Datasets.R
\name{age_model_zeeden}
\alias{age_model_zeeden}
\title{Age model of Zeeden et al., (2013) for the (154-174m) interval of the IODP 926 grey scale record}
\description{
Age model (anchor points) of the IODP 926 grey scale (154-174m) record of Zeeden et al., (2013) \cr
Anchored to the eccentricity-tilt-precession model p-0.5t of la 2004.
}
\details{
Column 1: Depth (meters)\cr
Column 2: Age (kyr)\cr
}
\references{
Christian Zeeden, Frederik Hilgen, Thomas Westerhold, Lucas Lourens, Ursula Röhl, Torsten Bickert,
Revised Miocene splice, astronomical tuning and calcareous plankton biochronology of ODP Site 926 between 5 and 14.4Ma,
Palaeogeography, Palaeoclimatology, Palaeoecology,Volume 369,2013,Pages 430-451,ISSN 0031-0182,
<doi:10.1016/j.palaeo.2012.11.009>


Laskar, Jacques, Philippe Robutel, Frédéric Joutel, Mickael Gastineau, Alexandre CM Correia,
 and Benjamin Levrard. "A long-term numerical solution for the insolation quantities of the Earth.
 " Astronomy & Astrophysics 428, no. 1 (2004): 261-285.
\url{https://www.aanda.org/articles/aa/pdf/2004/46/aa1335.pdf}
}
