% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve2time_unc.R
\name{curve2time_unc}
\alias{curve2time_unc}
\title{Convert the tracked curve to a depth time space with uncertainty}
\usage{
curve2time_unc(
  tracked_cycle_curve = NULL,
  tracked_cycle_period = NULL,
  tracked_cycle_period_unc = NULL,
  tracked_cycle_period_unc_dist = "n",
  n_simulations = NULL,
  output = 1
)
}
\arguments{
\item{tracked_cycle_curve}{Curve of the cycle tracked using the
\code{\link{track_period_wavelet}} function \cr
Any input (matrix or data frame) in which the first column is depth in
 meters and the second column is period in meters can be used.}

\item{tracked_cycle_period}{Period of the tracked curve in kyr.}

\item{tracked_cycle_period_unc}{uncertainty in the period of the tracked cycle}

\item{tracked_cycle_period_unc_dist}{distribution of the uncertainty of the
tracked cycle value need to be either "u" for uniform distribution or
"n" for normal distribution  \code{Default="n"}}

\item{n_simulations}{number of time series to be modeled}

\item{output}{If output = 1 a matrix with the predicted ages given the input for each run
is given. If output = 2 a matrix with 6 columns is generated,
the first column is depth/height, the other columns are the quantile
(0.025,0.373,0.5,0.6827,0.975) age values of the runs}
}
\value{
If output = 1 a matrix with the predicted ages given the input for each run
is given
If output = 2 a matrix with 6 columns is generated, the first column is
depth/height, the other columns are the quantile
(0.025,0.373,0.5,0.6827,0.975) age values of the runs
}
\description{
Converts the tracked curve to a depth time space while also
taking into account the uncertainty of the tracked astronomical cycle
}
\examples{
\donttest{
#Convert a tracked curve to a depth time space. The examples uses the
#magnetic susceptibility data set of Pas et al., (2018).

# perform the CWT
mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = FALSE,
omega_nr = 10)

#Track the 405 kyr eccentricity cycle in a wavelet spectra

#mag_track <- track_period_wavelet(astro_cycle = 405,
#                                   wavelet=mag_wt,
#                                   n.levels = 100,
#                                   periodlab = "Period (metres)",
#                                   x_lab = "depth (metres)")

#Instead of tracking, the tracked solution data set mag_track_solution is used
mag_track <- mag_track_solution

mag_track_complete <- completed_series(
  wavelet = mag_wt,
  tracked_curve = mag_track,
  period_up = 1.2,
  period_down = 0.8,
  extrapolate = TRUE,
  genplot = FALSE
)

# smooth the tracking of the 405 kyr eccentricity cycle
mag_track_complete <- loess_auto(time_series = mag_track_complete,
genplot = FALSE, print_span = FALSE)

#convert period in meters to sedrate depth vs time
mag_track_time<- curve2time_unc(tracked_cycle_curve = mag_track_complete,
tracked_cycle_period = 405,
tracked_cycle_period_unc = 2.4,
tracked_cycle_period_unc_dist = "n",
n_simulations = 25,
output=1 )

}
}
\references{
Routines for astrochronologic testing, astronomical time scale construction, and
time series analysis <\doi{doi:10.1016/j.earscirev.2018.11.015}>
}
\author{
Based on the \link[astrochron]{sedrate2time}
function of the 'astrochron' R package
}
