% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wavelet.R
\name{plot_wavelet}
\alias{plot_wavelet}
\title{Plots a wavelet power spectra}
\usage{
plot_wavelet(
  wavelet = NULL,
  plot.COI = TRUE,
  n.levels = 100,
  color.palette = "rainbow(n.levels, start = 0, end = 0.7)",
  useRaster = TRUE,
  periodlab = "Period (metres)",
  x_lab = "depth (metres)",
  keep_editable = FALSE
)
}
\arguments{
\item{wavelet}{wavelet object created using the \code{\link{analyze_wavelet}} function.}

\item{plot.COI}{Option to plot the cone of influence \code{Default=TRUE}.}

\item{n.levels}{Number of color levels \code{Default=100}.}

\item{color.palette}{Definition of color palette \code{Default="rainbow(n.levels, start = 0, end = 0.7)"}.}

\item{useRaster}{plot as a raster or vector image \code{Default=TRUE}.
WARNING plotting as a vector image is computationally intensive.}

\item{periodlab}{label for the y-axis \code{Default="Period (metres)"}.}

\item{x_lab}{label for the x-axis \code{Default="depth (metres)"}.}

\item{keep_editable}{Keep option to add extra features after plotting  \code{Default=FALSE}}
}
\value{
The output is a plot of a wavelet spectra.
}
\description{
Plot wavelet spectra using the outcome of the \code{\link{analyze_wavelet}} function.
}
\examples{
\donttest{
#Example 1. A plot of a wavelet spectra using the Total Solar Irradiance
# data set of Steinhilber et al., (2012)
TSI_wt <-
 analyze_wavelet(
   data = TSI,
   dj = 1/200,
   lowerPeriod = 16,
   upperPeriod = 8192,
   verbose = FALSE,
   omega_nr = 6
 )

plot_wavelet(
 wavelet = TSI_wt,
plot.COI = TRUE,
 n.levels = 100,
 color.palette = "rainbow(n.levels, start = 0, end = 0.7)",
 useRaster = TRUE,
 periodlab = "Period (years)",
x_lab = "years (before present)",
 keep_editable=FALSE
)

#Example 2. A plot of a wavelet spectra using the magnetic susceptibility
#data set of Pas et al., (2018)
mag_wt <-
analyze_wavelet(
data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = FALSE,
omega_nr = 10
)
plot_wavelet(
 wavelet = mag_wt,
 plot.COI = TRUE,
 n.levels = 100,
 color.palette = "rainbow(n.levels, start = 0, end = 0.7)",
 useRaster = TRUE,
 periodlab = "Period (metres)",
 x_lab = "depth (metres)",
  keep_editable=FALSE
)

#Example 3. A plot of a wavelet spectra using the greyscale
# data set of Zeeden et al., (2013)
grey_wt <-
 analyze_wavelet(
   data = grey,
   dj = 1/200,
   lowerPeriod = 0.02,
   upperPeriod = 256,
   verbose = FALSE,
   omega_nr = 8
 )
plot_wavelet(
 wavelet = grey_wt,
 plot.COI = TRUE,
 n.levels = 100,
 color.palette = "rainbow(n.levels, start = 0, end = 0.7)",
 useRaster = TRUE,
 periodlab = "Period (metres)",
 x_lab = "depth (metres)",
  keep_editable=FALSE
)



}
}
\references{
Angi Roesch and Harald Schmidbauer (2018). WaveletComp: Computational
Wavelet Analysis. R package version 1.1.
\url{https://CRAN.R-project.org/package=WaveletComp}

Gouhier TC, Grinsted A, Simko V (2021). R package biwavelet: Conduct Univariate and Bivariate Wavelet Analyses. (Version 0.20.21),
\url{https://github.com/tgouhier/biwavelet}

Torrence, C., and G. P. Compo. 1998. A Practical Guide to Wavelet Analysis.
Bulletin of the American Meteorological Society 79:61-78.
\url{https://paos.colorado.edu/research/wavelets/bams_79_01_0061.pdf}

Morlet, Jean, Georges Arens, Eliane Fourgeau, and Dominique Glard.
"Wave propagation and sampling theory—Part I: Complex signal and scattering in multilayered media.
" Geophysics 47, no. 2 (1982): 203-221.
\doi{<doi:10.1190/1.1441328>}

J. Morlet, G. Arens, E. Fourgeau, D. Giard;
Wave propagation and sampling theory; Part II, Sampling theory and complex waves.
 Geophysics 1982 47 (2): 222–236. \doi{<doi:10.1190/1.1441329>}
}
\author{
Code based on the \link[WaveletComp]{analyze.wavelet} and \link[WaveletComp]{wt.image} functions of the 'WaveletComp' R package
and \link[biwavelet]{wt} function of the 'biwavelet' R package which are based on the
wavelet MATLAB code written by Christopher Torrence and Gibert P. Compo (1998).
}
