% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_period_wavelet.R
\name{track_period_wavelet}
\alias{track_period_wavelet}
\title{Track the period of a cycle in a wavelet spectra}
\usage{
track_period_wavelet(
  astro_cycle = 405,
  wavelet = NULL,
  n.levels = 100,
  periodlab = "Period (metres)",
  x_lab = "depth (metres)",
  palette_name = "rainbow",
  color_brewer = "grDevices",
  plot_horizontal = TRUE
)
}
\arguments{
\item{astro_cycle}{Duration (in kyr) of the cycle which traced.}

\item{wavelet}{Wavelet object created using the \code{\link{analyze_wavelet}} function.}

\item{n.levels}{Number of color levels \code{Default=100}.}

\item{periodlab}{label for the y-axis \code{Default="Period (metres)"}.}

\item{x_lab}{label for the x-axis \code{Default="depth (metres)"}.}

\item{palette_name}{Name of the color palette which is used for plotting.
The color palettes than can be chosen depends on which the R package is specified in
the color_brewer parameter. The included R packages from which palettes can be chosen
from are; the 'RColorBrewer', 'grDevices', 'ColorRamps' and 'Viridis' R packages.
There are many options to choose from so please
read the documentation of these packages \code{Default=rainbow}.
The R package 'viridis' has the color palette options: “magma”, “plasma”,
“inferno”, “viridis”, “mako”, and “rocket”  and “turbo”
To see the color palette options of the The R pacakge 'RColorBrewer' run
the RColorBrewer::brewer.pal.info() function
The R package 'colorRamps' has the color palette options:"blue2green",
"blue2green2red", "blue2red",    "blue2yellow", "colorRamps",    "cyan2yellow",
"green2red", "magenta2green", "matlab.like", "matlab.like2" and    "ygobb"
The R package 'grDevices' has the built in  palette options:"rainbow",
"heat.colors", "terrain.colors","topo.colors" and "cm.colors"
To see even more color palette options of the The R pacakge 'grDevices' run
the grDevices::hcl.pals() function}

\item{color_brewer}{Name of the R package from which the color palette is chosen from.
The included R packages from which palettes can be chosen
are; the RColorBrewer, grDevices, ColorRamps and Viridis R packages.
There are many options to choose from so please
read the documentation of these packages. "\code{Default=grDevices}}

\item{plot_horizontal}{plot the wavelet horizontal or vertical eg y axis is depth or y axis power  \code{Default=TRUE}}
}
\value{
Results of the tracking of a cycle in the wavelet spectra is a matrix with 3 columns.
The first column is depth/time
The second column is the period of the tracked cycle
The third column is the sedimentation rate based on the duration (in time) of the tracked cycle
}
\description{
Interactively select points in a wavelet spectra to trace a period in a wavelet spectra.
The \code{\link{track_period_wavelet}} function plots a wavelet spectra in which spectral peaks can selected
allowing one to track a ridge hence one can track the a cycle with a changing period.
Tracking points can be selected in the Interactive interface and will be shown as white dots
when one wants to deselect a point the white dots can be re-clicked/re-selected and will turn red which
indicates that the previously selected point is deselected. Deselecting points can be quite tricky
due to the close spacing of  points and such the \code{\link{delpts_tracked_period_wt}} can be used to
delete points were previously selected using the \code{\link{track_period_wavelet}} function.
}
\examples{
\donttest{
#Track the 405kyr eccentricity cycle in the magnetic susceptibility record
# of the Sullivan core of Pas et al., (2018)

mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = FALSE,
omega_nr = 10)

mag_track <- track_period_wavelet(astro_cycle = 405,
                                  wavelet=mag_wt,
                                  n.levels = 100,
                                  periodlab = "Period (metres)",
                                  x_lab = "depth (metres)",
                                 palette_name = "rainbow",
                                 color_brewer="grDevices",
                                 plot_horizontal=TRUE)
}

}
\references{
Routines for astrochronologic testing, astronomical time scale construction, and
time series analysis <\doi{doi:10.1016/j.earscirev.2018.11.015}>
}
\author{
The function is based/inspired on the \link[astrochron]{traceFreq}
function of the 'astrochron' R package
}
