% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_uncertainty.R
\name{wavelet_uncertainty}
\alias{wavelet_uncertainty}
\title{Calculate the uncertainty associated with the wavelet analysis based on the Gabor uncertainty principle}
\usage{
wavelet_uncertainty(
  tracked_cycle = NULL,
  period_of_tracked_cycle = NULL,
  wavelet = NULL,
  multi = 1,
  verbose = TRUE,
  genplot_time = TRUE,
  genplot_uncertainty = TRUE,
  genplot_uncertainty_wt = TRUE
)
}
\arguments{
\item{tracked_cycle}{Curve of the cycle tracked using the \code{\link{track_period_wavelet}} function
Any input (matrix or dataframe) in which the first column is depth or
time and the second column is period should work}

\item{period_of_tracked_cycle}{period of the tracked curve (in kyr).}

\item{wavelet}{wavelet object created using the \code{\link{analyze_wavelet}} function.}

\item{multi}{multiple of the standard deviation to be used for defining uncertainty \code{Default=1}.}

\item{verbose}{Print text \code{Default=TRUE}.}

\item{genplot_time}{plot time curves with a upper and lower uncertainty based on Gabor uncertainty principle applied to the
continuous wavelet transform using a Mortlet wavelet, which uses which uses the omega number (number
of cycles in the wavelet) at one standard deviation to define the analytical uncertainty \code{Default=TRUE}}

\item{genplot_uncertainty}{Plot period curves with upper and lower uncertainty based on Gabor uncertainty principle applied to the
continuous wavelet transform using a Mortlet wavelet, which uses which uses the omega number
(number of cycles in the wavelet) to define uncertainty at one standard deviation \code{Default=TRUE}}

\item{genplot_uncertainty_wt}{generate a wavelet plot with the uncertainty based on Gabor uncertainty
principle applied to the continuous wavelet transform using a Mortlet wavelet superimposed on top of
original wavelet plot. The red curve is period of the tracked curve plus the analytical uncertainty.
The blue curve is period of the tracked curve min the analytical uncertainty.
The blakc black curve is the curve tracked using the '\code{Default=tracked_cycle_curve} function \code{Default=TRUE}}
}
\value{
Results pertaining to the uncertainty calculated based on the Gabor uncertainty principle.\cr
If the genplot_time is TRUE then a depth time plot will be plotted with 3 lines, the mean age,age plus
x times the standard deviation and age minus x times the standard deviation . \cr
If the genplot_uncertainty is TRUE then a curve will be plotted with the mean period, the tracked period plus
x times the standard deviation and the tracked period minus x times the standard deviation. \cr
If the genplot_uncertainty_wt is TRUE a wavelet spectra will be plotted with the tracked period, the tracked period plus
x times the standard deviation,the tracked period minus x times the standard devation and the area in between will be shaded in grey.\cr

Returns a matrix with 8 columns.\cr
The first column is called "depth" eg. depth \cr
The second column is "period" of the originally tracked period. \cr
The third column is "frequency" of the originally tracked period. \cr
The fourth column "uncertainty in frequency FWHM" is the uncertainty in frequency based on the Gabor uncertainty principle defined as
(FWHM) full width at half maximum. \cr
The fifth column "uncertainty in frequency x_times SD" is the uncertainty in frequency based on the Gabor uncertainty principle defined as
times x standard deviations. \cr
The sixth column "time mean" is the mean time based on the tracked period. \cr
The seventh column "time plus x_times sd" is the time based on the tracked period plus x times the standard deviation. \cr
The eight column "time min x_times sd" is the time based on the tracked period min x times the standard deviation. \cr
}
\description{
The \code{\link{wavelet_uncertainty}} function is used to calculate uncertainties associated
with the wavelet analysis based on the Gabor uncertainty principle applied to the
continuous wavelet transform using a Mortlet wavelet. The calculated uncertainty is the underlying
analytical uncertainty which is the result of applying the Gabor uncertainty principle to the
continuous wavelet transform using a Mortlet wavelet.
}
\examples{
\donttest{
#calculate the gabor uncertainty derived mathematical uncertainty of the \cr
#magnetic susceptibility record of the Sullivan core

mag_wt <- analyze_wavelet(data = mag,
dj = 1/100,
lowerPeriod = 0.1,
upperPeriod = 254,
verbose = TRUE,
omega_nr = 10)

#Track the 405 kyr eccentricity cycle in a wavelet spectra

#mag_track <- track_period_wavelet(astro_cycle = 405,
#                                   wavelet=mag_wt,
#                                   n.levels = 100,
#                                   periodlab = "Period (metres)",
#                                   x_lab = "depth (metres)")

#Instead of tracking, the tracked solution data set \code{\link{mag_track_solution}} is used \cr
mag_track <- mag_track_solution

mag_track_complete <- completed_series(
  wavelet = mag_wt,
  tracked_curve = mag_track,
  period_up = 1.2,
  period_down = 0.8,
  extrapolate = FALSE,
  genplot = FALSE
)

mag_track_complete <- loess_auto(time_series = mag_track_complete,
 genplot = TRUE, print_span = TRUE)

uncertainty <- wavelet_uncertainty(
  tracked_cycle = mag_track_complete,
  period_of_tracked_cycle = 405,
  wavelet = mag_wt,
  multi=1,
  verbose = TRUE,
  genplot_time = TRUE,
  genplot_uncertainty = TRUE,
  genplot_uncertainty_wt = TRUE
)
}
}
\references{
Gabor, Dennis. "Theory of communication. Part 1: The analysis of information."
Journal of the Institution of Electrical Engineers-part III: radio and
communication engineering 93, no. 26 (1946): 429-441.

Russell, Brian, and Jiajun Han. "Jean Morlet and the continuous wavelet transform.
" CREWES Res. Rep 28 (2016): 115.
}
